Apply addapted SF Patch #1908330
git-svn-id: http://svn.code.sf.net/p/xc3sprog/code/trunk@24 5e3be8be-922f-45f9-ad52-57b51d30232e

+++ jtag.h
@@ -1,5 +1,3 @@
-#include <stdio.h>
-
 /* JTAG routines
 
 Copyright (C) 2004 Andrew Rogers
@@ -73,7 +71,7 @@ class Jtag
     b[3]=(byte)((l>>24)&0xff);
   }
   inline unsigned long byteArrayToLong(byte *b){
-    return ((unsigned long)b[3]<<24)+((unsigned long)b[2]<<16)+((unsigned long)b[1]<<8)+(unsigned long)b[0];
+    return (b[3]<<24)+(b[2]<<16)+(b[1]<<8)+b[0];
   }
 };
